Allow cloudformation to create a file that holds the stack name and resource ID.
This is useful for deploys that don't necessarily know which stack name or resource ID they're dealing with; the instance now already knows itself and just reads it from a path.

Simply write a file `/home/stack-name` using cloudformation, containing the stack name. And pass it on in your `manifest` like:

```
            'cloudformation_signal': {
                'path:/home/stack-name': 'Asg'
            }
```