BaseHealthCheck retry-wrapper
This is the implementation from InstanceManager, with the predicate
extracted to an argument.
This means InstanceManager and ElbHealthCheck both get the `timeout`
function I rushed into ElbHealth.
Adding a `max_retries` flag, if you hate clocks.
Making `max_interval` a parameter, to limit the time between consecutive
checks.

Refactored `elb` child in manifest to accept a string (legacy fallback)
or an object. Said object still names the ELB, but provides a good hook
for `timeout`.